refactor: 카드 번호 로직 커스텀훅화
